Moved unit-tests from ColorPickerTest.ts to StylesTest.ts
